remove server from globals
