+++ nicegui/__init__.py
@@ -6,6 +6,7 @@ from .awaitable_response import AwaitableResponse
 from .client import Client
 from .logging import log
 from .nicegui import app
+from .server import Server
 from .tailwind import Tailwind
 from .version import __version__
 
@@ -20,6 +21,7 @@ __all__ = [
     'log',
     'optional_features',
     'run',
+    'Server',
     'Tailwind',
     'ui',
     '__version__',
